[maven-release-plugin]  copy for branch activemq-cpp-3.3.x
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/branches/activemq-cpp-3.3.x@1091396 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/test-benchmarks/benchmark/PerformanceTimer.h
@@ -63,7 +63,7 @@ namespace benchmark{
          * Gets the number of runs made so far
          * @returns unsigned int that counts the number of runs
          */
-        unsigned int getNumberOfRuns() const {
+        unsigned long long getNumberOfRuns() const {
             return numberOfRuns;
         }
